requirements.txt
autopep8==1.6.0
brotlipy
build
captum
certifi==2022.12.7
cffi
charset-normalizer
click
contextlib2==21.6.0
contourpy
cryptography
cycler
docker-pycreds==0.4.0
fonttools
gitdb==4.0.9
GitPython==3.1.27
idna
importlib-metadata
joblib==1.1.0
kiwisolver
matplotlib
munkres==1.1.4
numpy
packaging==21.3
pandas==2.0.0
pathtools==0.1.2
patsy
pep517
Pillow
pip-tools
platformdirs
pooch
promise==2.3
protobuf==3.19
psutil==5.9.1
pycodestyle==2.8.0
pycparser
pyOpenSSL
pyparsing
PySocks
python-dateutil
python-dotenv==0.20.0
pytorch-metric-learning==1.3.0
pytz
PyYAML==6.0
requests
schema==0.7.5
scikit-learn
scipy==1.10.1
seaborn
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
six
smac
smmap==5.0.0
statsmodels
threadpoolctl==3.1.0
toml==0.10.2
tomli
torch
torchaudio==0.11.0
torchvision
tqdm==4.64.0
typing_extensions
tzdata
umap-learn
unicodedata2
urllib3
wandb==0.12.16
zipp